appengine: Support either Device ID or an Alias in client and cmd.
In AppEngine API, Devices can be addressed through either their Device ID or
one of their Aliases. Reflect this by allowing to choose one of the two in the
client, or allow an heuristic to determine automatically what is which.

In astartectl, this is reflected by automatically determining whether the provided
identifier is an Astarte Device ID or an Alias. A --force-id-type flag is provided
to cover corner cases where the heuristic is ambiguous (e.g.: an alias which is still
a valid Astarte Device ID).

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>