client/appengine: Avoid flaky parsing when "value" is a path token
Datastream parsing was making too strong of an assumption of finding the bottom of a Datastream path
hierarchy whenever "value" was in the keys of a JSON object. However, if a path contained "value"
as one of the tokens, this would have caused the parser to believe that it had reached the last token
in the hierarchy blindly, whereas the parser had to dug deeper.

Whenever "value" is found in an object, check if it is actually a map[string]interface{} (aka a JSON
Object). In case it is, keep digging deeper. Also, port the same, more efficient, behavior to
Aggregate Datastreams (which already had this check on the timestamp field).

(Oh, also: hello Unit testing)

Fixes #48

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>