realm: Add the interfaces sync subcommand
sync should take away the tedious process of manually installing/updating large set
of interfaces. You can supply sync with any number of interface files, and it will
compute whether any of those interfaces should be installed or updated, depending on
the state of the realm, and perform those operations.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>