appengine/devices: Add send-data command
Long awaited, send-data finally allows southbound communications with the device.
When called on a server-owned interface, it sends data in the right format, taking
advantage of the most recent changes in astarte-go, which also ensures a decent level
of payload safety. This completes all main interactions with the device on astartectl
side.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>