device: Ensure send-data parses integer correctly for aggregates
Due to how JSON marshaling works in Go, all numbers are parsed as a float64. However,
when dealing with Astarte type validation, this caused failures for integer values due
to an inherent loss of precision upon conversion. Instead, if Math.trunc tells us we're
not losing precision, cast values to int to ensure that the type will be validated
(note that this works even if the integer is meant to represent a float, as there is
no precision loss involved).

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>